+++ apps/ai-dial-admin/src/components/EntityHeaderControls/Wrappers/ContainersButtonsWrapper.tsx
@@ -45,7 +45,6 @@ import {
 } from '@/src/utils/deployments/entity';
 import { getErrorNotification } from '@/src/utils/notification';
 import Delete from '@/src/components/EntityView/Modals/Delete/Delete';
-import DiscardModal from '@/src/components//EntityView/Modals/Discard/Discard';
 
 export interface ContainersButtonsWrapperProps {
   route: ApplicationRoute;
@@ -90,7 +89,6 @@ const ContainersButtonsWrapper: FC<ContainersButtonsWrapperProps> = ({
   const [buttonsClassNames, setButtonsClassNames] = useState('');
 
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isDiscardModalOpen, setIsDiscardModalOpen] = useState(false);
 
   const handleModalClose = useCallback(() => {
     setIsModalOpen(false);
@@ -173,21 +171,12 @@ const ContainersButtonsWrapper: FC<ContainersButtonsWrapperProps> = ({
     onDiscard?.();
   }, [dispatch, onDiscard]);
 
-  const onTryToDiscard = useCallback(() => {
-    setIsDiscardModalOpen(true);
-  }, []);
-
-  const onDiscardModalConfirm = useCallback(() => {
-    onStartDiscard();
-    setIsDiscardModalOpen(false);
-  }, [onStartDiscard]);
-
   return (
     <>
       <div className={containerClassNames}>
         {isChanged ? (
           <ChangedEntityButtons
-            onDiscard={onTryToDiscard}
+            onDiscard={onStartDiscard}
             onSave={onSave}
             disableSave={!isValid}
             saveLabel={t(isRedeployRequired ? ButtonsI18nKey.SaveAndRedeploy : ButtonsI18nKey.Save)}
@@ -289,13 +278,6 @@ const ContainersButtonsWrapper: FC<ContainersButtonsWrapperProps> = ({
           />,
           document.body,
         )}
-      {isDiscardModalOpen && (
-        <DiscardModal
-          onConfirm={onDiscardModalConfirm}
-          onClose={() => setIsDiscardModalOpen(false)}
-          onCancel={() => setIsDiscardModalOpen(false)}
-        />
-      )}
     </>
   );
 };
