+++ apps/ai-dial-admin/src/components/EntityHeaderControls/Wrappers/ImagesButtonsWrapper.tsx
@@ -33,7 +33,7 @@ import { getErrorNotification, getSuccessNotification } from '@/src/utils/notifi
 import { getUrnForEntity } from '@/src/utils/open-in-new-tab';
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
 import ImageDelete from '@/src/components/Deployments/Modals/ImageDelete';
-import DiscardModal from '@/src/components//EntityView/Modals/Discard/Discard';
+import ChangedEntityButtons from '@/src/components/EntityHeaderControls/Buttons/ChangedEntityButtons';
 
 export interface ImagesButtonsWrapperProps {
   image: Image;
@@ -74,7 +74,6 @@ const ImagesButtonsWrapper: FC<ImagesButtonsWrapperProps> = ({
   const [containerClassNames, setContainerClassNames] = useState(SELECT_ENTITY_HEADER_CLASS);
   const [buttonsClassNames, setButtonsClassNames] = useState('');
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isDiscardModalOpen, setIsDiscardModalOpen] = useState(false);
 
   const allowEditing = useMemo(() => {
     return image.buildStatus !== IMAGE_STATUS.BUILT && image.buildStatus !== IMAGE_STATUS.BUILDING;
@@ -176,39 +175,24 @@ const ImagesButtonsWrapper: FC<ImagesButtonsWrapperProps> = ({
     setButtonsClassNames(classNames((isTablet || isMobile) && SELECT_ENTITY_MOBILE_HEADER_BUTTONS_CLASS));
   }, [isTablet, isMobile]);
 
-  const onTryToDiscard = useCallback(() => {
-    setIsDiscardModalOpen(true);
-  }, []);
-
-  const onDiscardModalConfirm = useCallback(() => {
-    onDiscard?.();
-    setIsDiscardModalOpen(false);
-  }, [onDiscard]);
-
   return (
     <>
       <div className={containerClassNames}>
         {isChanged ? (
           <div className="flex flex-row gap-3 w-full p-3 lg:p-0">
-            <DialNeutralButton
-              className={buttonsClassNames}
-              label={t(ButtonsI18nKey.Discard)}
-              onClick={onTryToDiscard}
-            />
-            <DialPrimaryButton
-              className={buttonsClassNames}
-              label={t(ButtonsI18nKey.SaveAsNewVersion)}
-              onClick={onOpenSaveNewVersionModal}
-              disabled={isDisableSave}
-            />
-            {allowEditing && !forceNewVersion && (
+            <ChangedEntityButtons
+              disableSave={isDisableSave}
+              onDiscard={onDiscard}
+              onSave={onSave}
+              isSaveAllowed={allowEditing && !forceNewVersion}
+            >
               <DialPrimaryButton
                 className={buttonsClassNames}
-                label={t(ButtonsI18nKey.Save)}
-                onClick={onSave}
+                label={t(ButtonsI18nKey.SaveAsNewVersion)}
+                onClick={onOpenSaveNewVersionModal}
                 disabled={isDisableSave}
               />
-            )}
+            </ChangedEntityButtons>
           </div>
         ) : (
           <div className="flex flex-row items-center w-full gap-x-4">
@@ -327,13 +311,6 @@ const ImagesButtonsWrapper: FC<ImagesButtonsWrapperProps> = ({
           />,
           document.body,
         )}
-      {isDiscardModalOpen && (
-        <DiscardModal
-          onConfirm={onDiscardModalConfirm}
-          onClose={() => setIsDiscardModalOpen(false)}
-          onCancel={() => setIsDiscardModalOpen(false)}
-        />
-      )}
     </>
   );
 };
